+++ CHANGELOG.md
@@ -24,7 +24,10 @@ Adding a requirement of a major version of a dependency is breaking a contract.
 Dropping a requirement of a major version of a dependency is a new contract.
 
 ## [Unreleased]
-[Unreleased]: https://bitbucket.org/atlassian/infrastructure/branches/compare/master%0Drelease-2.1.1
+[Unreleased]: https://bitbucket.org/atlassian/infrastructure/branches/compare/master%0Drelease-2.2.0
+
+## [2.2.0] - 2018-09-21
+[2.2.0]: https://bitbucket.org/atlassian/infrastructure/branches/compare/release-2.2.0%0Drelease-2.1.1
 
 ### Added
 - Add the `CPU_UTILISATION` dimension.